Reorganize directories (sub-directory components for vlite)
