+++ examples/audio/index.html
@@ -10,7 +10,7 @@
 		<audio id="player-html5" class="vlite-js" src="https://cdn.plyr.io/static/demo/Kishi_Bashi_-_It_All_Began_With_a_Burst.mp3"></audio>
 		<script src="../../dist/vlite.js"></script>
 		<script>
-			new Vlitejs('#player-html5', {
+			const player = new Vlitejs('#player-html5', {
 				options: {
 					autoplay: false,
 					controls: true,
@@ -22,12 +22,12 @@
 				onReady: (player) => {
 					console.log(player)
 
-					player.container.addEventListener('play', () => console.log('play'))
-					player.container.addEventListener('pause', () => console.log('pause'))
-					player.container.addEventListener('ended', () => console.log('ended'))
-					player.container.addEventListener('progress', () => console.log('progress'))
-					player.container.addEventListener('volumechange', () => console.log('volumechange'))
-					player.container.addEventListener('timeupdate', () => console.log('timeupdate'))
+					player.on('play', () => console.log('play'))
+					player.on('pause', () => console.log('pause'))
+					player.on('ended', () => console.log('ended'))
+					player.on('progress', () => console.log('progress'))
+					player.on('volumechange', () => console.log('volumechange'))
+					player.on('timeupdate', () => console.log('timeupdate'))
 				}
 			})
 			window.player = player
