+++ src/vlite/assets/styles/player.css
