+++ src/vlite/components/big-play/assets/styles/big-play.css
