+++ src/vlite/components/big-play/config.js
