+++ src/vlite/components/control-bar/assets/scripts/control-bar.ts
