+++ src/vlite/components/control-bar/assets/scripts/templates/control-bar.ts
