+++ src/vlite/components/control-bar/assets/styles/control-bar.css
