+++ src/vlite/components/control-bar/config.js
