+++ src/vlite/components/loader/assets/scripts/loader.ts
