+++ src/vlite/components/loader/config.js
