+++ src/vlite/components/overlay/config.js
