+++ src/vlite/components/poster/assets/scripts/poster.ts
