+++ src/vlite/components/poster/assets/styles/poster.css
