+++ src/vlite/components/poster/config.js
